Fix silent integrity failure by enforcing introspection-based hashing (breaking change).
- Modified `compute_audit_hash` in `src/coreason_manifest/utils/audit.py` to always use introspection (v2 logic).
- Removed legacy `V1_FIELDS` and `hash_algorithm` support.
- Updated `tests/test_audit_integrity.py` to verify secure behavior and removed legacy tests.
- Linting fixes applied to `src/coreason_manifest/utils/audit.py`.
